% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePolygonSamplePoints.R
\name{computeCellSize}
\alias{computeCellSize}
\title{computeCellSize -- Compute the \code{cellsize} parameter for \code{spsample} for a sample area}
\usage{
computeCellSize(area = 6000, aunits = "acre", pattern = "hexagonal")
}
\arguments{
\item{area}{numeric: Area represented by each sample point in acres or hectares.}

\item{aunits}{character string: Units for \code{area}. Valid choices are "acres" and
"hectares". Only the first letter is needed and case doesn't matter.}

\item{pattern}{character string: the desired sampling pattern. Valid choices are "regular"
and "hexagonal". Only the first letter is needed and cae doesn't matter.}
}
\value{
Numeric value used as input to the \code{spsample} function for
the \code{cellsize} parameter. Units will match the units associated
with \code{aunits} (feet for acres and meters for hectares).
}
\description{
Compute the value for the \code{cellsize} parameter for the \code{spsample} function in the
\code{sp} package to generate a sample grid representing \code{area} square units. This is
a helper function for \code{generatePolygonSamplePoints} to help develop sample points
that represent a given sample area.
}
\examples{
# 6000 acre sample area with value returned in feet
computeCellSize(3000, "acre")
# 1000 hectare sample area with value returned in meters
computeCellSize(1000, "hectare")
}
