% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WASnoCo}
\alias{WASnoCo}
\title{Polygon for Snohomish County Washington}
\format{
A sf data object with 5 attributes and 1 geometry:
\describe{
\item{STATEFP}{FIPS code for Washington state}
\item{COUNTYFP}{FIPS code for Snohomish County}
\item{NAME}{County name}
\item{HECTARES}{Area of Snohomish County in hectares}
\item{ACRES}{Area of Snohomish County in acres}
}
}
\usage{
WASnoCo
}
\description{
A polygon representing Snohomish County Washington. Polygon
has been projected to UTM zone 10N (EPSG: 26910).
}
\details{
Orignal data was taken from US Census Tiger data retrieved
using the tigris package.
}
\keyword{datasets}
